Finales setup and continuous deployment test.
The Readme.md file was updated to explain better this project.
The project descripton was added under the header component and the css
was adjusted to achieve a style improvement.
If the added descripton is under header in the live version of the
project, it means that the master branch of this repository is
synchronized with the hosting which means that the continuous deployment
of the project is correctly configured. 😄